+++ packages/core/src/canvas/renderer/retained-backing/types.ts
@@ -1,0 +1,35 @@
+import type { Image as CKImage, Surface } from 'canvaskit-wasm'
+
+import type { SceneGraph } from '@open-pencil/scene-graph'
+
+export interface RenderContentVersion {
+  pageId: string | null
+  sceneVersion: number
+  positionPreviewVersion: number
+  fontGeneration: number
+}
+
+export interface SceneBackingGeometry {
+  panX: number
+  panY: number
+  zoom: number
+  width: number
+  height: number
+  dpr: number
+  worldX: number
+  worldY: number
+  worldWidth: number
+  worldHeight: number
+}
+
+export interface SceneBacking extends RenderContentVersion, SceneBackingGeometry {
+  image: CKImage
+}
+
+export interface SceneBackingBuild extends RenderContentVersion, SceneBackingGeometry {
+  surface: Surface
+  graph: SceneGraph
+  childIds: string[]
+  index: number
+  startedAt: number
+}
