+++ packages/docs/.vitepress/sidebars.ts
@@ -114,6 +114,7 @@ export const developmentSidebar = (
         ? [
             { text: 'Roadmap', link: '/development/roadmap' },
             { text: 'Navigation Performance', link: '/development/navigation-performance' },
+            { text: 'Renderer Lifecycle', link: '/development/renderer-lifecycle' },
             { text: 'Renderer Profiler', link: '/development/renderer-profiler' },
             { text: 'Vector Conversion', link: '/development/vector-conversion' },
           ]
